Fix importer onboarding E2E fixtures
